Fix cleaning logic in deploy script
Summary:
The sequence `rm .git/index; git reset --hard` was intended to remove
all files in the work tree. But it actually was effectively a no-op: it
cleared the index, and then reset to the previous commit, not the empty
index. The behavior was the same for the initial commit, because the
parent deploy was empty, but for subsequent deploys this caused a
failure due to the previous build not having been removed.

Test Plan:
Run `./deploy.sh -n`, which fails before this commit and passes after
it.

wchargin-branch: fix-cleaning